+++ ecoz2py/ecoz2_extension_build.py
@@ -18,7 +18,7 @@ ffibuilder.set_source(
     "../ecoz2/src/include"
   ],
   sources=[
-    '../ecoz2/src/ecoz/ecoz2.c',
+    '../ecoz2/src/ecoz2/ecoz2.c',
     '../ecoz2/src/utl/utl.c',
     '../ecoz2/src/utl/fileutil.c',
     '../ecoz2/src/utl/list.c',
@@ -49,6 +49,7 @@ ffibuilder.set_source(
     '../ecoz2/src/hmm/hmm_adjustb.c',
     '../ecoz2/src/hmm/hmm_file.c',
     '../ecoz2/src/hmm/hmm_refinement.c',
+    '../ecoz2/src/hmm/hmm_prob.c',
     '../ecoz2/src/hmm/hmm_log_prob.c',
     '../ecoz2/src/hmm/hmm_genQopt.c',
     '../ecoz2/src/hmm/hmm_estimateB.c',